fix compile

+++ src/llama-batch.cpp
@@ -406,7 +406,7 @@ struct llama_batch_ext * llama_batch_ext_init_from_embd(
 }
 
 int32_t llama_batch_ext_set_pos(struct llama_batch_ext * batch, llama_pos * pos, size_t n_pos) {
-    if (batch->n_tokens != n_pos) {
+    if ((size_t) batch->n_tokens != n_pos) {
         return -1;
     }
     memcpy(batch->pos, pos, n_pos * sizeof(llama_pos));
